Task3

+++ Tyuiu.KurbanovFA.Sprint2.Task3.V22/Program.cs
@@ -27,7 +27,7 @@ namespace Tyuiu.KurbanovFA.Sprint2.Task3.V22
             Console.WriteLine("***************************************************************************");
             Console.WriteLine("*                                                                         *");
             Console.WriteLine("Введите значение X: ");
-            int x = Convert.ToInt32(Console.ReadLine());
+            double x = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("*                                                                         *");
             Console.WriteLine("***************************************************************************");
             Console.WriteLine("* РЕЗУЛЬТАТ:                                                              *");
